#
Added image URL handling and Turkish error messages

Added the `ImageUrl` property to various classes and implemented image uploading functionality in services. This allows for the inclusion and updating of an image URL when creating or updating a city, country, event, or venue. Also, error messages were translated from English to Turkish. In the UI, elements were added to display and upload images. The `Ticket.API` project was removed from the solution and the `Notification.API` project was updated. Lastly, the `ImageUrl` property was added to event-related classes and used to update the event's image URL.